+++ src/main/java/app/visa/repository/PasseportRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.visa.entity.Passeport;
 
-public interface PasseportRepository extends JpaRepository<Passeport, Long> {
+public interface PasseportRepository extends JpaRepository<Passeport, Integer> {
 }
