+++ src/main/java/app/visa/repository/SituationFamilialeRepository.java
@@ -3,5 +3,5 @@ package app.visa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import app.visa.entity.SituationFamiliale;
 
-public interface SituationFamilialeRepository extends JpaRepository<SituationFamiliale, Long> {
+public interface SituationFamilialeRepository extends JpaRepository<SituationFamiliale, Integer> {
 }
